requirements.txt
anyio==3.6.1
asgiref==3.5.2
astroid==2.12.12
attrs==21.4.0
autopep8==1.6.0
Brotli==1.0.9
certifi==2022.5.18.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
ConfigArgParse==1.5.3
coverage==6.5.0
Cython==0.29.32
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
exceptiongroup==1.0.0
fastapi==0.78.0
filelock==3.8.0
flake8==6.0.0
flake8-html==0.4.2
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
genbadge==1.1.0
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
h11==0.13.0
httpretty==1.1.4
identify==2.5.8
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
locust==2.15.1
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.5
nodeenv==1.7.0
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
psutil==5.9.4
pycodestyle==2.10.0
pydantic==1.9.1
pyflakes==3.0.1
Pygments==2.13.0
pylint==2.15.5
pyparsing==3.0.9
pytesseract==0.3.10
pytest==7.2.0
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
pyzmq==25.0.2
requests==2.27.1
roundrobin==0.0.4
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.16.6
Werkzeug==2.2.3
wrapt==1.14.1
zope.event==4.6
zope.interface==6.0
